requirements.txt
paho-mqtt
pvporcupine
vosk
piper-tts
pyaudio
soundfile
numpy
